+++ lib/generators/ductwork/update/update_generator.rb
@@ -86,6 +86,11 @@ module Ductwork
         migration_template "db/add_process_id_to_ductwork_executions.rb",
                            "db/migrate/add_process_id_to_ductwork_executions.rb"
       end
+
+      if Ductwork::Process.column_names.exclude?("role")
+        migration_template "db/add_role_to_ductwork_processes.rb",
+                           "db/migrate/add_role_to_ductwork_processes.rb"
+      end
     end
   end
 end